A cleanup for "MDEV-11952 Oracle-style packages: Phase#1: ..."
Fixed that a syntax error inside CREATE PACKAGE [BODY] caused a crash.

The problem happened because thd->lex was not restored to the top level LEX
when recovering from a syntax error in LEX::cleanup_lex_after_parse_error().
thd->lex was left assigned to the local sublex corresponding to a routine
definition.